get tokenizer config
