+++ scripts/evaluator/evaluate_utils/__init__.py
@@ -1,3 +1,4 @@
 from .formatter import *
 from .metrics import *
 from .prompt import *
+from .to_be_deprecated import *
