+++ scripts/llm_inference_adapter.py
@@ -1,6 +1,8 @@
 import os
 from config_singleton import WandbConfigSingleton
 from langchain_community.chat_models import ChatOpenAI
+from utils import download_tokenizer_config
+
 
 def get_llm_inference_engine():
     instance = WandbConfigSingleton.get_instance()
@@ -10,6 +12,7 @@ def get_llm_inference_engine():
     if api_type == "vllm":
         # vLLMサーバーを起動
         from vllm_server import start_vllm_server
+
         start_vllm_server()
 
         # LangChainのVLLMインテグレーションを使用
@@ -20,10 +23,10 @@ def get_llm_inference_engine():
             **cfg.generator,
         )
 
-        # TODO 別のtokenizerを使っているときにも対応できるようにする。from vllm import LLM, SamplingParamsを使えば良さそう
-        from transformers import AutoTokenizer
-        tokenizer = AutoTokenizer.from_pretrained(cfg.model.pretrained_model_name_or_path)
-        cfg.update({"special_tokens_map": tokenizer.special_tokens_map})
+        tokenizer_config = download_tokenizer_config(
+            repo_id=cfg.model.pretrained_model_name_or_path
+        )
+        cfg.update({"tokenizer_config": tokenizer_config})
 
         return llm
